Add setResizePolicy/getResizePolicy functions
The RingBuffer now holds a flag that determines the action to take when
resizing the buffer to a size smaller than the current size. The
previous version of RingBuffer (and the now default behavior) preserved
the front of the buffer while truncating the back end. The behavior can
be changed to preserve the back end of the buffer by using the new
setter function.